requirements.txt
# Core ML Framework
torch>=2.0.0
torchvision>=0.15.0

# Image Processing
Pillow>=9.0.0

# Web Application
streamlit>=1.25.0

# Data Visualization
matplotlib>=3.5.0
seaborn>=0.11.0

# Data Manipulation
numpy>=1.21.0
pandas>=1.3.0

# Optional: For better performance
# tensorboard>=2.10.0  # Uncomment if you want TensorBoard logging
# wandb>=0.13.0       # Uncomment if you want Weights & Biases integration
